Working Furnace Steak, balance changes
Enchantments for the Steak Bucket work properly, the bucket can be eaten when on low durability but it gives no nutrition to the player. The furnace steak smelts up to half a random itemstack using fuel from the player's inventory instantly but inefficiently.
Steak Golem is now weaker overall (still no sounds or textures)